\name{case0801}
\alias{case0801}
\docType{data}
\title{Island Area and Number of Species}
\description{
  The data are the numbers of reptile and amphibian species and the
  island areas for seven islands in the West Indies.
}
\usage{case0801}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{\code{Area}}{area of island (in square miles)}
    \item{\code{Species}}{number of reptile and amphibian species on island}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\examples{
str(case0801)
plot(case0801)
}
\keyword{datasets}
