\name{ex0218}
\alias{ex0218}
\docType{data}
\title{Peter and Rosemary Grant's Finch Beak Data}
\description{ 
  In the 1980s, biologists Peter and Rosemary Grant caught and measured
  all the birds from more than 20 generations of finches on the
  Galapagos island of Daphne Major. In one of those years, 1977, a
  severe drought caused vegetation to wither, and the only remaining
  food source was a large, tough seed, which the finches ordinarily
  ignored. Were the birds with larger and stronger beaks for opening
  these tough seeds more likely to survive that year, and did they tend
  to pass this characteristic to their offspring? The data are beak
  depths (height of the beak at its base) of 751 finches caught the year
  before the drought (1976) and 89 finches captured the year after the
  drought (1978).
}
\usage{ex0218}
\format{
   A data frame with 840 observations on the following 2 variables.
  \describe{
    \item{Year}{Year the finch was caught, 1976 or 1978}
    \item{Depth}{Beak depth of the finch (mm)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Grant, P. (1986). \emph{Ecology and Evolution of Darwin's Finches},
  Princeton University Press, Princeton, N.J. 
}
\seealso{
  \code{\link{case0201}}
}
\examples{
str(ex0218)
}
\keyword{datasets}
