\name{ex0222}
\alias{ex0222}
\docType{data}
\title{Cholesterol in Urban and Rural Guatemalans}
\description{
  This data comes from an observational study to contrast cholesterol
  levels in rural and urban Guatemalan Indians
}
\usage{ex0222}
\format{
  A data frame with 94 observations on the following 2 variables.
  \describe{
    \item{\code{Cholesterol}}{Serum total cholesterol of individual (in mg/l)}
    \item{\code{Group}}{a factor with levels \code{"Rural"} and \code{"Urban"}
      indicating to which group the individual belongs}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Tejada, C., Charm, S., Guzman, M., Mendez, J. and Kurland,
  G. (1964). The Blood Viscosity of Various Socioeconomic Groups in
  Guatemala, \emph{American Journal of Clinical Nutrition}: 303--308.
}
\examples{
str(ex0222)
}
\keyword{datasets}
