\name{ex0331}
\alias{ex0331}
\docType{data}
\title{Iron Supplementation}
\description{
  A randomized experiment was performed on mice to determine whether two
  forms of iron are retained differently.  If one type is retained
  especially well it may be more useful as a dietary supplement for
  humans.
}
\usage{ex0331}
\format{
  A data frame with 36 observations on the following 2 variables.
  \describe{
    \item{\code{Iron}}{percentage of iron retained in each mouse}
    \item{\code{Supplement}}{factor variable with levels \code{"Fe3"}
      and \code{"Fe4"}}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Rice, J. (1987).  \emph{Mathematical Statistics and Data Analysis},
  Wadsworth. 
}
\examples{
str(ex0331)
}
\keyword{datasets}
