\name{ex0332}
\alias{ex0332}
\docType{data}
\title{College Tuition}
\description{
  Tuition in dollars of 20 private and 20 public U.S. colleges and
  universities for 1993--1994.
}
\usage{ex0332}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{Private}}{tuition in dollars of 20 private schools}
    \item{\code{PubIn}}{tuition in dollars of 20 public schools
      (in-state tuition)} 
    \item{\code{PubOut}}{tuition in dollars of 20 public schools
      (out-of-state tuition)} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  \emph{1995 U.S. News and World Report's Guide to America's Best
    Colleges.}
}
\examples{
str(ex0332)
}
\keyword{datasets}
