\name{ex0724}
\alias{ex0724}
\docType{data}
\title{Crab Claw Size and Force}
\description{
  As part of a study of the effects of predatory intertidal crab species
  on snail populations, researchers measured the mean closing forces and
  the propdus heights of the claws on several crabs of three species.
}
\usage{ex0724}
\format{
  A data frame with 38 observations on the following 3 variables.
  \describe{
    \item{\code{Force}}{closing strength of claw of the crab}
    \item{\code{Height}}{propodus height of claw of the crab}
    \item{\code{Species}}{species to which the crab belongs}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Yamada, S.B. and Boulding, E.G. (1992).  Shell--breaking Efficiency of
  Predatory Crabs Influences the Distribution of an Intertidal Snail,
  Technical Report, Zoology Department, Oregon State University.
}
\examples{
str(ex0724)
}
\keyword{datasets}
