\name{ex0728}
\alias{ex0728}
\docType{data}
\title{Number of Stories and Building Height}
\description{
  The \emph{1994 World Almanac} reports heights and number of stories
  for notable tall buildings in North America.  The data in this data
  frame  are a random sample of size 60 of those for which dates of
  completion were available.
}
\usage{ex0728}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{year of completion}
    \item{\code{Height}}{height of building}
    \item{\code{Stories}}{number of stories of building}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\examples{
str(ex0728)
}
\keyword{datasets}
