\name{ex0816}
\alias{ex0816}
\docType{data}
\title{Meat Processing}
\description{
  The data in \code{\link{case0702}} are a subset of the complete data on
  postmortum pH in 12 steer carcasses.
}
\usage{ex0816}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{Time}}{time after slaughter (hours)}
    \item{\code{pH}}{pH level in postmortem muscle}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Schwenke, J.R. and Milliken, G.A. (1991).  On the Calibration Problem
  Extended to Nonlinear Models, \emph{Biometrics} \bold{47}(2): 563--574.
}
\seealso{
  \code{\link{case0702}}
}
\examples{
str(ex0816)
}
\keyword{datasets}
