\name{ex0817}
\alias{ex0817}
\docType{data}
\title{Biological Pest Control}
\description{
  In a study of the effectiveness of biological control of the exotic
  weed tansy ragwort, researchers manipulated the exposure to the
  ragwort flea beetle on 15 plots that had been planted with a high
  density of ragwort.  Harvesting the plots the next season, they
  measured the average dry mass of ragwort remaining (grams/plant) and
  the flea beetle load (beetles/gram of ragwort dry mass) to see if the
  ragwort plants in plots with high flea beetle loads were smaller as a
  result of herbivory by the beetles.
}
\usage{ex0817}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Load}}{flee beetle load (in beetles/gram of ragwort dry mass)}
    \item{\code{Mass}}{dry mass of ragwort weed}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  McEvoy, P. and Cox, C. (1991).  Successful Biological Control of
  Ragwort, \emph{Senecio jacobaea}, by introducing insects in Oregon,
  \emph{Ecological Applications} \bold{1}(4): 430--442.
}
\examples{
str(ex0817)
}
\keyword{datasets}
