\name{ex0818}
\alias{ex0818}
\docType{data}
\title{Chernobyl Fallout}
\description{
  One of the most dangerous contaminants deposited over European
  countries following the Chernobyl accident in April 1986 was
  radioactive cesium.  To study cesium transfer from contaminated soil
  to plants, researchers collected soil samples and samples of mushroom
  mycelia from 17 wooded locations in Umbria, Central Italy, from August
  1986 to November 1989.  The data are measured concentrations (Bq/kg)
  of cesium in the soil and in the mushrooms.
}
\usage{ex0818}
\format{
  A data frame with 17 observations on the following 2 variables.
  \describe{
    \item{\code{Mushroom}}{Cesium concentrations in mushrooms (in Bq/kg)}
    \item{\code{Soil}}{Cesium concentrations in soil (in Bq/kg)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Borio, R. et al. (1991).  Uptake of Radiocesium by Mushrooms,
  \emph{Science of the Total Environment} \bold{106}: 183--190.
}
\examples{
str(ex0818)
}
\keyword{datasets}
