\name{ex0822}
\alias{ex0822}
\docType{data}
\title{Ecosystem Decay}
\description{
  Data are the number of butterfly species in 16 islands of forest of
  various sizes in otherwise cleared areas in Brazil.
}
\usage{ex0822}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Area}}{area (ha) of forest patch}
    \item{\code{Species}}{number of butterfly species}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Lovejoy, T.E., Rankin, J.M., Bierregaard, Jr., R.O., Brown, Jr., K.S.,
  Emmons, L.H.  and  van der Voort, M. (1984). Ecosystem decay of Amazon
  forest remnants \emph{in} Nitecki, M.H. (ed.) \emph{Extinctions},
  University of Chicago Press. 
}
\examples{
str(ex0822)
}
\keyword{datasets}
