\name{ex0825}
\alias{ex0825}
\docType{data}
\title{The Dramatic U.S. Presidential Election of 2000}
\description{
  Data set shows the number of votes for Buchanan and Bush in all 67
  counties in Florida during the U.S. presidential election of November
  7, 2000.
}
\usage{ex0825}
\format{
  A data frame with 67 observations on the following 3 variables.
  \describe{
    \item{\code{County}}{a character vector indicating the county}
    \item{\code{Buchanan2000}}{votes cast for P. Buchanan}
    \item{\code{Bush2000}}{votes cast for G.W. Bush}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\seealso{
  \code{\link{ex1222}}
}
\examples{
str(ex0825)
}
\keyword{datasets}
