\name{ex0914}
\alias{ex0914}
\docType{data}
\title{Pace of Life and Heart Disease}
\description{
  In four regions of the US (Northeast, Midwest, South and West), in
  three different sized metropolitan regions, researchers measured
  indicators of pace of life.
}
\usage{ex0914}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Bank}}{bank clerk speed}
    \item{\code{Walk}}{pedestrian walking speed}
    \item{\code{Talk}}{postal clerk talking speed}
    \item{\code{Heart}}{age adjusted death rate due to heart disease}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Levine, R.V. (1990). The Pace of Life, \emph{American Scientist}
  \bold{78}: 450--459.
}
\examples{
str(ex0914)
}
\keyword{datasets}
