\name{ex1026}
\alias{ex1026}
\docType{data}
\encoding{latin1}
\title{Thinning of Ozone Layer}
\description{
  Depletion of the ozone layer allows the most damaging ultraviolet
  radiation to reach the Earth's surface.  To measure the relationship,
  researchers sampled the ocean column at various depths at 17 locations
  around Antarctica during the austral spring of 1990.
}
\usage{ex1026}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{Inhibit}{percent inhibition of primary phytoplankton
      production in water} 
    \item{UVB}{UVB exposure}
    \item{Surface}{a factor with levels \code{"Deep"} and
      \code{"Surface"} }
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Smith, R.C., Prézelin, B.B., Baker, K.S., Bidigare, R.R., Boucher,
  N.P., Coley, T., Karentz, D., MacIntyre, S., Matlick, H.A., Menzies,
  D., Ondrusek, M., Wan, Z. and Waters, K.J. (1992).  Ozone Depletion:
  Ultraviolet Radiation and Phytoplankton Biology in Antarctic Waters,
  \emph{Science} \bold{255}(5047): 952--959. 
}
\examples{
str(ex1026)
}
\keyword{datasets}
