\name{ex1509}
\alias{ex1509}
\docType{data}
\title{Sunspot Counts for 1749--1948}
\description{
  Time series data set of annual counts of sunspots.
}
\usage{ex1509}
\format{
  A data frame with 200 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{Spots}}{number of sunspots}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Waldmeier, M. (1961). \emph{The Sunspot Activity in the Years
    1610--1960}, Federal Observatory, Zurich.
}
\examples{
str(ex1509)
}
\keyword{datasets}
