\name{ex1515}
\alias{ex1515}
\docType{data}
\title{S\&P 500}
\description{
  Data on the value of a \eqn{\$}{$}1 U.S. stock investment in 1871 at
  the end of each year, based on the Standard and Poor (S\eqn{\&}{&}P)
  500 Composite stock index.
}
\usage{ex1515}
\format{
  A data frame with 129 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{SPReturn}}{S\eqn{\&}{&}P composite stock index (\eqn{\$}{$})}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\examples{
str(ex1515)
}
\keyword{datasets}
