\name{ex1605}
\alias{ex1605}
\docType{data}
\title{Nature---Nurture}
\description{
  Data are a subset from an observational, longitudinal, study on
  adopted children.  Is child's intelligence related to intelligence of
  the biological mother and the intelligence of the adoptive mother?
}
\usage{ex1605}
\format{
  A data frame with 62 observations on the following 6 variables.
  \describe{
    \item{\code{AMED}}{adoptive mother's years of education}
    \item{\code{BMIQ}}{biological mother's score on IQ test}
    \item{\code{Age2IQ}}{IQ of child at age 2}
    \item{\code{Age4IQ}}{IQ of child at age 4}
    \item{\code{Age8IQ}}{IQ of child at age 8}
    \item{\code{Age13IQ}}{IQ of child at age 13}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Skodak, M. and Skeels, H.M. (1949). A Final Follow-up Study of One
  Hundred Adopted Children, \emph{Journal of Genetic Psychology}
  \bold{75}: 85--125.
}
\seealso{
  \code{\link{ex1319}}
}
\examples{
str(ex1605)
}
\keyword{datasets}
