\name{ex1613}
\alias{ex1613}
\docType{data}
\title{Flea Beetle Distinction}
\description{
  Data are the measurements from two very similar species of flea beetle.
}
\usage{ex1613}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{Jnt1}}{measurement of first joint in micrometers}
    \item{\code{Jnt2}}{measurement of second joint in micrometers}
    \item{\code{Species}}{a factor with levels \code{"conc"} and
      \code{"heik"}} 
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Lubischew, A.A. (1962).  On the Use of Discriminant Functions in
  Taxonomy, \emph{Biometrics} \bold{18}: 455--477.
}
\examples{
str(ex1613)
}
\keyword{datasets}
