\name{ex1708}
\alias{ex1708}
\docType{data}
\title{Pig Fat}
\description{
  Actual pig fat and measurements of pig fat from magnetic resonance
  images at 13 locations for 12 pigs.
}
\usage{ex1708}
\format{
  A data frame with 12 observations on the following 14 variables.
  \describe{
    \item{\code{Fat}}{actual pig fat (in percent)}
    \item{\code{M1}}{magnetic resonance image at location 1}
    \item{\code{M2}}{magnetic resonance image at location 2}
    \item{\code{M3}}{magnetic resonance image at location 3}
    \item{\code{M4}}{magnetic resonance image at location 4}
    \item{\code{M5}}{magnetic resonance image at location 5}
    \item{\code{M6}}{magnetic resonance image at location 6}
    \item{\code{M7}}{magnetic resonance image at location 7}
    \item{\code{M8}}{magnetic resonance image at location 8}
    \item{\code{M9}}{magnetic resonance image at location 9}
    \item{\code{M10}}{magnetic resonance image at location 10}
    \item{\code{M11}}{magnetic resonance image at location 11}
    \item{\code{M12}}{magnetic resonance image at location 12}
    \item{\code{M13}}{magnetic resonance image at location 13}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Glasbey, C.A and Flowler, P.A. (1992).  Regression Models Fitted Using
  Conditional Independence to Estimate Pig Fatness from Magnetic
  Resonance Images, \emph{The Statistician} \bold{41}(2): 179--184.
}
\examples{
str(ex1708)
}
\keyword{datasets}
