\name{ex1713}
\alias{ex1713}
\docType{data}
\title{Church Distinctiveness}
\description{
  Data show measures that differ among denominations of American
  Protestant and Catholic churches.
}
\usage{ex1713}
\format{
  A data frame with 18 observations on the following 6 variables.
  \describe{
    \item{\code{Denomination}}{a character vector indicating the
      church denomination} 
    \item{\code{Distinct}}{distinctiveness (strictness of discipline on
      a seven point scale) } 
    \item{\code{Attend}}{average percentage of weeks that individuals
      attended a church meeting (\% weekly)}
    \item{\code{NonChurch}}{average number of secular organisations to
      which members belong}
    \item{\code{StrongPct}}{average percentage of members that describe
      themselves as being strong church members (\%)} 
    \item{\code{AnnInc}}{average income of members (US\eqn{\$}{$})}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Iannaccone, L.R. (1994).  Why Strict Churches Are Strong,
  \emph{American Journal of Sociology} \bold{99}(5): 1180--1211.
}
\examples{
str(ex1713)
}
\keyword{datasets}
