\name{ex2011}
\alias{ex2011}
\docType{data}
\title{Space Shuttle}
\description{
  This data frame contains the launch temperatures (degrees Fahrenheit)
  and an indicator of O-ring failures for 24 space shuttle launches
  prior to the space shuttle \emph{Challenger} disaster of January 28,
  1986. 
}
\usage{ex2011}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Temp}}{Launch temperature (in degrees Fahrenheit)} 
    \item{\code{Failure}}{Indicator of O-ring failure}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\seealso{
 \code{\link{case0401}}, 
 \code{\link{ex2223}} 
}
\examples{
str(ex2011)
}
\keyword{datasets}
