\name{ex2225}
\alias{ex2225}
\docType{data}
\title{Body Size and Reproductive Success in a Population of Male Bullfrogs}
\description{
  As an example of field observation in evidence of theories of sexual
  selection, S.J. Arnold and M.J. Wade presented the following data set
  on size and number of mates observed in 38 bullfrogs.
}
\usage{ex2225}
\format{
  A data frame with 38 observations on the following 2 variables.
  \describe{
    \item{\code{Bodysize}}{Body size (in mm)}
    \item{\code{Mates}}{Number of mates}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Arnold, S.J. and Wade, M.J. (1984). On the Measurement of Natural and
  Sexual Selection: Aplications, \emph{Evolution} \bold{38}: 720--734.
}
\examples{
str(ex2225)
}
\keyword{datasets}
