\name{Sleuth2-package}
\alias{Sleuth2-package}
\alias{Sleuth2}
\docType{package}
\title{
The R Sleuth2 package
}
\description{
Data sets from Ramsey and Schafer's "Statistical Sleuth (2nd ed)"
}
\details{This package contains a variety of datasets.  For a complete
  list, use \code{library(help="Sleuth2")} or
  \code{Sleuth2Manual()}.
}
\author{
Original by F.L. Ramsey and D.W. Schafer

Modifications by Daniel W Schafer, Jeannie Sifneos and Berwin A Turlach

Maintainer: Berwin A Turlach \email{Berwin.Turlach@gmail.com}
}
\keyword{package}
