\name{Sleuth2Manual}
\alias{Sleuth2Manual}
\title{
  Manual of the R Sleuth2 package
}
\description{
  If the option \dQuote{pdfviewer} is set, this command will display the
  PDF version of the help pages.  
}
\usage{
Sleuth2Manual()
}
\references{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\author{
  Berwin A Turlach \email{Berwin.Turlach@gmail.com}
}
\examples{
\dontrun{Sleuth2Manual()}
}
\keyword{documentation}
