\name{Sleuth3-package}
\alias{Sleuth3-package}
\alias{Sleuth3}
\docType{package}
\title{
The R Sleuth3 package
}
\description{
Data sets from Ramsey and Schafer's "Statistical Sleuth (3rd ed)"
}
\details{This package contains a variety of datasets.  For a complete
  list, use \code{library(help="Sleuth3")} or
  \code{Sleuth3Manual()}.
}
\author{
Original by F.L. Ramsey and D.W. Schafer

Modifications by Daniel W Schafer, Jeannie Sifneos and Berwin A Turlach

Maintainer: Berwin A Turlach \email{Berwin.Turlach@gmail.com}
}
\keyword{package}
