\name{case2001}
\alias{case2001}
\docType{data}
\title{Survival in the Donner Party}
\description{
  This data frame contains the ages and sexes of the adult (over 15
  years) survivors and nonsurvivors of the Donner party.
}
\usage{case2001}
\format{
  A data frame with 45 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{Age of person}
    \item{\code{Sex}}{Sex of person}
    \item{\code{Status}}{Whether the person survived or died}
  }
}
\details{
  In 1846 the Donner and Reed families left Springfield, Illinois, for
  California by covered wagon.  In July, the Donner Party, as it became
  known, reached Fort Bridger, Wyoming.  There its leaders decided to
  attempt a new and untested rote to the Sacramento Valley.  Having
  reached its full size of 87 people and 20 wagons, the party was
  delayed by a difficult crossing of the Wasatch Range and again in the
  crossing of the desert west of the Great Salt Lake.  The group became
  stranded in the eastern Sierra Nevada mountains when the region was
  hit by heavy snows in late October.  By the time the last survivor was
  rescued on April 21, 1847, 40 of the 87 members had died from famine
  and exposure to extreme cold.
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Grayson, D.K. (1990). Donner Party Deaths: A Demographic Assessment,
  \emph{Journal of Anthropological Research} \bold{46}: 223--242.
}
\seealso{
  \code{\link{ex1918}}
}
\examples{
str(case2001)
}
\keyword{datasets}
