\name{ex0428}
\alias{ex0428}
\docType{data}
\title{Darwin's Data}
\description{
  Plant heights (inches) for 15 pairs of plants of the same age, one of
  which was grown from a seed from a cross-fertilized flower and the
  other of which was grown from a seed from a self-fertilized flower.
}
\usage{ex0428}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Cross}}{height (inches) of cross-fertilized plant}
    \item{\code{Self}}{height (inches) of self-fertilized plant}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Andrews, D.F. and Herzberg, A.M. (1985).  \emph{Data: A Collection of
  Problems from many Fields for the Student and Research Worker},
  Springer-Verlag. 
}
\examples{
str(ex0428)
}
\keyword{datasets}
