\name{ex0727}
\alias{ex0727}
\docType{data}
\encoding{latin1}
\title{Male Displays}
\description{
  Black wheatears are small birds in Spain and Morocco.  Males of the
  species demonstrate an exaggerated sexual display by carrying many
  heavy stones to nesting cavities.  This 35--gram bird transports, on
  average, 3.1 kg of stones per nesting season!  Different males carry
  somewhat different sized stones, prompting a study on whether larger
  stones may be a signal of higher health status.  Soler et
  al. calculated the average stone mass (g) carried by each of 21 male
  black wheatears, along with T-cell response measurements reflecting
  their immune systems' strengths. 
}
\usage{ex0727}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{Mass}{average mass of stones carried by bird (in g)}
    \item{Tcell}{T-cell response measurement (in mm)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
  Soler, M., Martín-Vivaldi, M., Marín J. and Møller, A. (1999). Weight
  lifting and health status in the black wheatears, \emph{Behavioral
  Ecology} \bold{10}(3): 281--286.
}
\examples{
str(ex0727)
}
\keyword{datasets}
