\name{ex0915}
\alias{ex0915}
\docType{data}
\title{Rainfall and Corn Yield}
\description{
  Data on corn yield and rainfall in six U.S. corn--producing states
  (Iowa, Nebraska, Illinois, Indiana, Missouri and Ohio), recorded for
  each year from 1890 to 1927.
}
\usage{ex0915}
\format{
  
  \describe{
    \item{\code{Year}}{year of observation (1890--1927)}
    \item{\code{Yield}}{average corn yield for the six states (in bu/acre)}
    \item{\code{Rainfall}}{average rainfall in the six states (in in/year)}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  Ezekiel, M. and Fox, K.A. (1959). \emph{Methods of Correlation and
    Regression Analysis}, John Wiley \& Sons, New York.
}
\examples{
str(ex0915)
}
\keyword{datasets}
