\name{ex1509}
\alias{ex1509}
\docType{data}
\title{Sunspots}
\description{
  The data are the annual sunspot counts in each year from 1700 to 2010.
}
\usage{ex1507}
\format{
   A data frame with 311 observations on the following 2 variables.
  \describe{
    \item{Year}{year in which sunspots were counted,
      from 1700 to 2010} 
    \item{Sunspots}{the number of sunspots observed in a year} 
  }
}
\source{
   Ramsey, F.L. and Schafer, D.W. (2013). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (3rd ed)}, Cengage Learning.
}
\references{
 SIDC-Solar Influences Dta Center, \url{http://sidc.oma.be/sunspot-data/} 
(July 15, 2011). 
}
\examples{
str(ex1509)
}
\keyword{datasets}
